refactor: reorganize user form hooks and schemas into a new directory structure
